Moved all remaining files from 'code/unotui/' to 'code/unotui/utility'
