+++ code/unotui/CMakeLists.txt
@@ -1,12 +1,3 @@
-target_sources_local(
-  ${APP_LIBRARY_NAME}
-  PUBLIC
-    coordinates.h
-    font_symbols.h
-    types.h
-  PRIVATE
-    types.cpp
-)
 
 add_subdirectory( entities )
 add_subdirectory( systems )
