+++ code/unotui/entities/widgets/geometry/texture.h
@@ -5,7 +5,7 @@
 
 #include <vector>
 
-#include <unotui\types.h>
+#include <unotui\utility\types.h>
 
 struct texture_geometry : geometry
 {
