+++ code/unotui/entities/widgets/w_editabletext.cpp
@@ -9,7 +9,7 @@
 #include <unotui\utility\widget.h>
 
 #include <unotui\entities\events\widget_event.h>
-#include <unotui\font_symbols.h>
+#include <unotui\utility\font_symbols.h>
 
 #include <cmath> // Remainder.
