+++ code/unotui/utility/general.cpp
@@ -7,7 +7,7 @@
 
 #include <unotui\utility\deps\opengl_includes.h>
 
-#include <unotui\coordinates.h> // For coordinates. Come on. (Unit, point.)
+#include <unotui\utility\coordinates.h> // For coordinates. Come on. (Unit, point.)
 
 #include <unotui\entities\ent_window.h>
 #include <unotui\entities\ent_opengl.h>
