+++ code/unotui/utility/general.h
@@ -7,7 +7,7 @@
 
 #include <unotui\utility\deps\opengl_includes.h>
 
-#include <unotui\coordinates.h> // For coordinates. Come on. (Unit, point.)
+#include <unotui\utility\coordinates.h> // For coordinates. Come on. (Unit, point.)
 
         //:: Textures.
 unsigned int loadTexture( const uint8_t* Data, int Length, GLint internal_format, GLenum format, const std::string& = "NO_NAME", bool mipmaps = true );
